api init
