+++ api/Cargo.toml
@@ -8,10 +8,15 @@ repository ="https://github.com/thor314/zkhn-rust-api"
 version    ="0.1.0"
 
 [dependencies]
+models = { path = "../models" }
+
 ammonia="3.3.0"
 anyhow="1.0"
 axum={ version="0.7.3", features=["macros"] }
 chrono={ version="0.4.34", features=["serde"] }
+diesel-async={ version="0.4.1", features=["postgres", "deadpool"] }
+diesel={ version="2.1.4", features=["postgres", "uuid", "chrono"] }
+diesel-derive-enum={ version="2.1.0", features=["postgres"] }
 regex="1.10.3"
 scrypt = "0.11.0"
 shuttle-axum="0.39.0"
@@ -33,12 +38,8 @@ url="2.5.0"
 # axum-extra = { version = "0.9.2", features = ["cookie", "typed-header"] }
 # tower-http = { version = "0.5.1", features = ["fs"] }
 
-[dev-dependencies]
-rstest           ="0.18"                                 # testing convenience proc-macros
-arbitrary        ={ version="1.1", features=["derive"] } # see quickcheck
-quickcheck       ="1.0"                                  # lowkey proptest suite
-quickcheck_macros="1"                                    # derive 
-
-[[test]]
-name="integration"
-path="tests/tests.rs"
+# [dev-dependencies]
+# rstest           ="0.18"                                 # testing convenience proc-macros
+# arbitrary        ={ version="1.1", features=["derive"] } # see quickcheck
+# quickcheck       ="1.0"                                  # lowkey proptest suite
+# quickcheck_macros="1"                                    # derive
