+++ api/src/api/items.rs
@@ -3,10 +3,10 @@ use diesel::prelude::*;
 use diesel_async::{AsyncPgConnection, RunQueryDsl};
 use uuid::Uuid as Uid;
 
-use crate::schema::{items, items::dsl::items as items_dsl};
-use crate::{
-  error::MyError,
-  // models::comment::{Comment, NewCommentPayload},
-  // schema::{comments, comments::dsl::comments as comments_dsl},
-  SharedState,
-};
+// use crate::schema::{items, items::dsl::items as items_dsl};
+// use crate::{
+//   error::MyError,
+//   // models::comment::{Comment, NewCommentPayload},
+//   // schema::{comments, comments::dsl::comments as comments_dsl},
+//   SharedState,
+// };
