--- api/src/models/mod.rs
@@ -1,7 +1,0 @@
-pub mod comment;
-pub mod item;
-pub mod moderation_log;
-pub mod user;
-pub mod user_favorite;
-pub mod user_hidden;
-pub mod user_vote;
