--- api/src/models/user_favorite.rs
@@ -1,21 +1,0 @@
-use axum::{extract::State, response::IntoResponse};
-use chrono::{DateTime, NaiveDate, NaiveDateTime, Utc};
-use diesel::{prelude::*, sql_types::*, QueryDsl, Queryable, Selectable, SelectableHelper};
-use diesel_async::{AsyncConnection, AsyncPgConnection, RunQueryDsl};
-use serde::{Deserialize, Serialize};
-use uuid::Uuid as Uid;
-
-use crate::schema::user_favorites;
-
-#[derive(Queryable, Selectable, Debug, Serialize, Deserialize)]
-// match to a schema for selectable
-#[diesel(table_name = user_favorites)]
-// use postgres, improve compiler error messages.
-#[diesel(check_for_backend(diesel::pg::Pg))]
-/// Represents a user's favorite item, including type and timestamp.
-pub struct UserFavorite {
-  pub username:  String,
-  pub item_type: String,
-  pub item_id:   Uid,
-  pub date:      NaiveDateTime,
-}
